Updated all hidKeys...() function to store u64 instead of u32. (#28)
Also fixed some whitespace in usbds example

+++ fs/sdmc/source/main.c
@@ -38,7 +38,7 @@ int main(int argc, char **argv)
         hidScanInput();
 
         //hidKeysDown returns information about which buttons have been just pressed (and they weren't in the previous frame)
-        u32 kDown = hidKeysDown(CONTROLLER_P1_AUTO);
+        u64 kDown = hidKeysDown(CONTROLLER_P1_AUTO);
 
         if (kDown & KEY_PLUS) break; // break in order to return to hbmenu
